+++ threedigrid_builder/tests/test_dict_out.py
@@ -52,7 +52,7 @@ def test_meta(dict_out):
         ("lines", "node_2", 2),  # reference to node
         ("lines", "cross_id1", 4),  # reference to cross section def id, not increased
         ("breaches", "id", 1),
-        ("breaches", "line_id", 5),  # reference to line
+        ("breaches", "line_id", 4),  # reference to line
     ],
 )
 def test_not_off_by_one(dict_out, group, dataset, expected):